Cafeteria menu
This project retrieves and displays daily menu information from the Compass Group website. A Python script is provided to automate this process and display the menu on the screen at scheduled times. The script is designed to run automatically according to a specified time. The project aims to integrate seamlessly with Magic Mirror, providing users with up-to-date menu information. Please note that the restaurant must have a JSON link on the website for the script to function properly.